Bot/requirements.txt
httpx==0.24.1
python-telegram-bot==20.4
python-dotenv==1.0.0
asyncio==3.4.3